# documentazione e database/openapi.yaml
openapi: 3.0.3
info:
  title: Fooody REST API
  description: API RESTful per la gestione delle ordinazioni e delle consegne a domicilio dell'attività Fooody.
  version: 1.0.0
  contact:
    name: Maikol Gaparroni
    email: maikol.gasparroni@student.univaq.it
servers:
  - url: http://localhost:8080/Fooody/api
    description: Server locale (Apache Tomcat)
paths:
...
  /ordini:
    get:
      security:
        - bearerAuth: [] # Questo endpoint richiede autenticazione con token JWT
      summary: Elenco ordini con filtri.
      tags: [Ordini]
      parameters:
        - in: query
          name: dataInserimento
          schema:
            type: string
            format: date
        - in: query
          name: stato
          schema:
            type: string
            enum: ["inserito", "in preparazione", "pronto", "in consegna", "consegnato", "annullato"]
      responses:
        '200':
          description: Elenco degli ordini restituito con successo.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Ordine'
    post:
      security:
        - bearerAuth: []
      summary: Creazione di un nuovo ordine completo
      description: Inserisce l'ordine intero comprensivo di prodotti e caratteristiche, ricalcolando automaticamente totale e tempistiche.
      tags: [Ordini]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NuovoOrdineRequest'
      responses:
        '201':
          description: Ordine creato con successo.
        '400':
          description: Impossibile creare un ordine vuoto.
        '500':
          description: Errore interno del server durante la creazione dell'ordine.
  /ordini/{idOrdine}:
    delete:
      security:
        - bearerAuth: [] # Questo endpoint richiede autenticazione con token JWT
      summary: Annulla ordine
      tags:
        - Ordini
      parameters:
        - in: path
          name: idOrdine
          required: true
          schema:
            type: integer
          description: ID dell'ordine da annullare.
      responses:
        '204':
          description: Ordine annullato con successo.
        '404':
          description: Ordine non trovato.
  /ordini/{idOrdine}/dettagli:
    post:
      security:
        - bearerAuth: [] # Questo endpoint richiede autenticazione con token JWT
      summary: Insermento prodotto in ordine
      tags:
        - Ordini
      parameters:
        - in: path
          name: idOrdine
          required: true
          schema:
            type: integer
          description: ID dell'ordine a cui aggiungere il prodotto.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NuovoDettaglioRequest'
      responses:
        '201':
          description: Prodotto aggiunto all'ordine con successo.
    get:
      security:
        - bearerAuth: [] # Questo endpoint richiede autenticazione con token JWT
      summary: Elenco prodotti in un ordine
      tags:
        - Ordini
      parameters:
        - in: path
          name: idOrdine
          required: true
          schema:
            type: integer
          description: ID dell'ordine di cui ottenere i dettagli.
      responses:
        '200':
          description: Elenco dei prodotti nell'ordine restituito con successo.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DettaglioProdotto'
  /ordini/{idOrdine}/tempo-stimato:
    get:
      security:
        - bearerAuth: [] # Questo endpoint richiede autenticazione con token JWT
      summary: Tempo stimato per la consegna.
      tags:
        - Ordini
      parameters:
        - in: path
          name: idOrdine
          required: true
          schema:
            type: integer
          description: ID dell'ordine di cui ottenere il tempo stimato per la consegna.
      responses:
        '200':
          description: Tempo stimato per la consegna restituito con successo.
          content:
            application/json:
              schema:
                type: object
                properties:
                  tempoStimatoMinuti:
                    type: integer
                    example: 30
                  messaggio:
                    type: string
                    example: "Il tempo stimato per la consegna è di circa 30 minuti."
  /ordini/{idOrdine}/prezzo-totale:
    get:
      security:
        - bearerAuth: [] # Questo endpoint richiede autenticazione con token JWT
      summary: Prezzo totale dell'ordine.
      tags:
        - Ordini
      parameters:
        - in: path
          name: idOrdine
          required: true
          schema:
            type: integer
          description: ID dell'ordine di cui ottenere il prezzo totale.
      responses:
        '200':
          description: Prezzo totale dell'ordine restituito con successo.
          content:
            application/json:
              schema:
                type: object
                properties:
                  prezzoTotale:
                    type: number
                    format: float
                    example: 25.50
                  messaggio:
                    type: string
                    example: "Il prezzo totale dell'ordine è di 25.50 euro."
  /ordini/{idOrdine}/stati:
    post:
      security:
        - bearerAuth: [] # Questo endpoint richiede autenticazione con token JWT
      summary: Aggiorna lo stato di un ordine
      tags:
        - Ordini
      parameters:
        - in: path
          name: idOrdine
          required: true
          schema:
            type: integer
          description: ID dell'ordine di cui aggiornare lo stato.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nuovoStato:
                  type: string
                  enum: ["in preparazione", "pronto", "in consegna", "consegnato", "annullato"]
                  example: "in preparazione"
      responses:
        '200':
          description: Stato dell'ordine aggiornato con successo.
  /ordini/{idOrdine}/operatori:
    get:
      security:
        - bearerAuth: [] # Questo endpoint richiede autenticazione con token JWT
      summary: Elenco operatori coinvolti in un ordine
      tags:
        - Ordini
      parameters:
        - in: path
          name: idOrdine
          required: true
          schema:
            type: integer
          description: ID dell'ordine di cui ottenere l'elenco degli operatori coinvolti.
      responses:
        '200':
          description: Elenco degli operatori coinvolti nell'ordine restituito con successo.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    nome: { type: string, example: "Mario Rossi"}
                    cognome: { type: string, example: "Rossi"}

  #-------------------
  # Personale
  #-------------------
...
components:
  schemas:
...
    Ordine:
      type: object
      properties:
        idOrdine:
          type: integer
          example: 1
        timeInserimento:
          type: string
          format: date-time
          example: "2024-06-01T12:30:00Z"
        OrarioConsegnaRichiesto:
          type: string
          format: date-time
          example: "2024-06-01T13:30:00Z"
        statoCorrente:
          type: string
          enum: ["inserito", "in preparazione", "pronto", "in consegna", "consegnato", "annullato"]
          example: "in preparazione"
        prezzoTotale:
          type: number
          format: float
          example: 25.50
    #############################################
    NuovoOrdineRequest:
      type: object
      properties:
        idCliente:
          type: integer
          example: 1
        orarioConsegnaRichiesto:
          type: string
          format: date-time
          example: "2026-06-29 20:30:00"
        items:
          type: array
          items:
            $ref: '#/components/schemas/NuovoDettaglioRequest'
    #############################################
    NuovoDettaglioRequest:
      type: object
      properties:
        idProdotto:
          type: integer
          example: 1
        quantita:
          type: integer
          example: 2
        caratteristicheScelte:
          type: array
          items:
            type: integer
          example: [1, 3]
    #############################################
    DettaglioProdotto:
      type: object
      properties:
        idDettaglio:
          type: integer
          example: 1
        nomeProdotto:
          type: string
          example: "Pizza Margherita"
        quantita:
          type: integer
          example: 2
        caratteristiche:
          type: array
          items:
            type: string
          example: ["Senza zucchero", "Con panna"]
        tempoPreparazione:
          type: integer
          description: Tempo di preparazione stimato in minuti per questo prodotto con le caratteristiche scelte
          example: 20
        procedura:
          type: string
          description: Procedura di preparazione del prodotto necessaria alla cucina, comprensiva delle caratteristiche scelte.
          example: "1. Stendere la pasta. 2. Aggiungere il pomodoro. 3. Aggiungere la mozzarella. 4. Cuocere in forno per 20 minuti."
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Utilizza un token JWT per autenticarsi nelle richieste alle risorse protette. Il token deve essere incluso nell'header Authorization con il prefisso "Bearer".